+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Develop
+-----
+- Fix wiring for ``@classmethod``.
+
 4.1.5
 -----
 - Fix Travis CI windows and MacOS builds.